tests: move FalseToNullConverter tests and InvalidDateConverter tests into appropriate test folder
